==> scripts/02_richness-maps.Rmd <==
---
title: "02_richness-maps"
author: "Kate Sheridan"
date: "9/15/2022"
output: html_document
---

This script generates maps of species richness from the data we downloaded from OBIS and GBIF.

```{r setup, include=FALSE}
#install packages if needed
#devtools::install_github("azizka/speciesgeocodeR")
#install.packages('raster')

library(speciesgeocodeR)
library(raster)
library(tidyverse)
library(ggplot2)
library(here)
```

When we load in the files, we'll split them between the two classes to analyze separately.

```{r load-in}
poly <- read_csv(here('data', 'rawdata',
                            '20220909_obis-gbif_bivalve-polychaete-subset.csv')) %>%
  filter(class == "Polychaeta")

biv <- read_csv(here('data', 'rawdata',
                            '20220909_obis-gbif_bivalve-polychaete-subset.csv')) %>%
  filter(class == "Bivalvia")

# world map
world <- map_data("world")
```



# Polychaetes


```{r polychaetes-map}
# fix to work with RichnessGrid
poly2 <- poly %>%
  # trim to study range
  filter(decimal_latitude > 30 & decimal_longitude < -116.4) %>%
  filter(decimal_latitude < 63 & decimal_longitude > -142.5) %>%
  # rename so richness grid understands
  rename(decimalLatitude = decimal_latitude,
         decimalLongitude = decimal_longitude) %>%
  select(worms_sciname, decimalLatitude, decimalLongitude) %>% 
  distinct()

# use RichnessGrid to calculate species richness per grid square 
poly_alpha <- RichnessGrid(poly2, type = "spnum")

## convert to plot
poly_alphaplot <- data.frame(rasterToPoints(poly_alpha))

# Visualize
ggplot() + geom_map(data = world, 
                    map = world, 
                    aes(x = long, y = lat, map_id = region), 
                    fill = "grey80") +
  xlim(min(poly2$decimalLongitude, na.rm = T),  
       max(poly2$decimalLongitude, na.rm = T)) + 
  ylim(min(poly2$decimalLatitude, na.rm = T), 
      max(poly2$decimalLatitude, na.rm = T)) + 
  # note that somehow lat and long got switched
  geom_raster(data = poly_alphaplot, aes(x = y, y = x, fill = layer)) + 
  coord_fixed() + 
  theme_bw() + 
  labs(title = "Number of Polychaete Species", fill = "") + 
  theme(axis.title = element_blank())

ggsave(here('output', 'maps', '20220924_polychaete_richness.png'))
```


# Bivalves

```{r polychaetes-map}
#fix to work with RichnessGrid
biv2 <- biv %>%
    # trim to study range
  filter(decimal_latitude > 30 & decimal_longitude < -116.4) %>%
  filter(decimal_latitude < 63 & decimal_longitude > -142.5) %>%
  rename(decimalLatitude = decimal_latitude,
         decimalLongitude = decimal_longitude) %>%
  select(worms_sciname, decimalLatitude, decimalLongitude) %>% 
  distinct()

# RichnessGrid to calculate number of species x square
biv_alpha <- RichnessGrid(biv2, type = "spnum")

# convert to plot
biv_alphaplot <- data.frame(rasterToPoints(biv_alpha))

# Visualize
ggplot() + 
  geom_map(data = world, 
                    map = world, 
           aes(x = long, y = lat, map_id = region), 
           fill = "grey80") +
  # limit plot area to data
  xlim(min(biv2$decimalLongitude, na.rm = T),  
       max(biv2$decimalLongitude, na.rm = T)) + 
  ylim(min(biv2$decimalLatitude, na.rm = T), 
      max(biv2$decimalLatitude, na.rm = T)) + 
  # coordinates get flipped somehow
  geom_raster(data = biv_alphaplot, 
              aes(x = y, y = x, fill = layer)) + 
  coord_fixed() +
  theme_bw() + 
  labs(title = "Number of Bivalve Species", fill = "") + 
  theme(axis.title = element_blank())

ggsave(here('output', 'maps', '20220924_bivalve_richness.png'))
```
Quantify data

```{r}
## should check by genus and species
length(unique(biv2$worms_sciname)) #143
```
